+++ rust/tests/mailbox.rs
@@ -1,17 +1,16 @@
 #[allow(dead_code)]
 mod constants;
 mod contracts;
-mod event;
 mod validator;
 
-use contracts::strk::mailbox::{mailbox, Bytes};
+use contracts::strk::mailbox::{mailbox, Bytes, Dispatch as DispatchEvent};
 use ethers::{
     prelude::parse_log, providers::Middleware, signers::Signer, types::TransactionReceipt,
 };
-use event::parse_dispatch_from_res;
 use starknet::{
     accounts::{Account, ConnectedAccount},
-    core::types::{FieldElement, MaybePendingTransactionReceipt},
+    core::types::{Event, FieldElement, MaybePendingTransactionReceipt},
+    core::utils::get_selector_from_name,
     providers::{AnyProvider, Provider},
 };
 
@@ -20,6 +19,19 @@ use crate::{
     contracts::{eth, strk},
     validator::TestValidators,
 };
+
+/// Parse the dispatch event from the receipt events
+pub fn parse_dispatch_from_res(events: &[Event]) -> DispatchEvent {
+    let key = get_selector_from_name("Dispatch").unwrap(); // safe to unwrap
+    let found = events.iter().find(|v| v.keys.contains(&key)).unwrap();
+
+    DispatchEvent {
+        sender: cainome::cairo_serde::ContractAddress(found.data[0]),
+        destination_domain: found.data[1].try_into().unwrap(),
+        recipient_address: cainome::cairo_serde::ContractAddress(found.data[2]),
+        message: (found.data[3], found.data[4]).try_into().unwrap(),
+    }
+}
 
 async fn send_msg_strk_to_evm<M, S>(
     from: &strk::Env,
